# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Textile ERP API - نظام إدارة رولات القماش
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in tenant
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Production Orders
...
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseRequest"
      responses:
        "201":
          description: Created warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /warehouses/{id}:
    get:
      operationId: getWarehouse
      tags: [warehouses]
      summary: Get warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWarehouseRequest"
      responses:
        "200":
          description: Updated warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"

  # Warehouse movements
  /warehouse-movements:
    get:
      operationId: listWarehouseMovements
      tags: [warehouses]
      summary: List warehouse movements
      parameters:
        - name: fabricRollId
          in: query
          schema:
            type: integer
        - name: warehouseId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of warehouse movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WarehouseMovement"
    post:
      operationId: createWarehouseMovement
      tags: [warehouses]
      summary: Move roll between warehouses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseMovementRequest"
      responses:
        "201":
          description: Created movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseMovement"

  # Customers
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - email
        - fullName
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
      required:
        - email
        - password
        - fullName
        - role
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        password:
          type: string
...
    Warehouse:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        location:
          type: string
        capacity:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - name
        - location
        - isActive
        - createdAt
        - updatedAt
    CreateWarehouseRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
      required:
        - name
        - location
    UpdateWarehouseRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        isActive:
          type: boolean
    WarehouseMovement:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        fabricRollId:
          type: integer
        fromWarehouseId:
          type: ["integer", "null"]
        toWarehouseId:
          type: ["integer", "null"]
        movedById:
          type: integer
        reason:
          type: ["string", "null"]
        movedAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - fabricRollId
        - movedById
        - movedAt
        - createdAt
    CreateWarehouseMovementRequest:
      type: object
      properties:
        fabricRollId:
          type: integer
        fromWarehouseId:
          type: integer
        toWarehouseId:
          type: integer
        reason:
          type: string
      required:
        - fabricRollId
        - toWarehouseId
...